requirements.txt
google-genai>=1.73.0
httpx>=0.28.0
python-dotenv>=1.0.0
pytest>=7.0.0
